move to shared utils
